WIP supporting more than one split
